Models, views and viewmodels separated in theire own packages
