+++ app/src/main/java/com/example/cardgame/viewmodels/GameplayViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.cardgame.viewmodels
+
+import androidx.lifecycle.ViewModel
+
+class GameplayViewModel : ViewModel() {
+
+}
