+++ app/src/main/java/com/example/cardgame/views/MainFragment.kt
@@ -1,7 +1,6 @@
-package com.example.cardgame
+package com.example.cardgame.views
 
 import android.content.Context
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -9,6 +8,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import androidx.fragment.app.Fragment
+import com.example.cardgame.R
 import com.example.cardgame.utils.Consts
 
 class MainFragment : Fragment() {
@@ -26,7 +26,8 @@ class MainFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        val rootView = inflater.inflate(R.layout.fragment_main, container,
+        val rootView = inflater.inflate(
+            R.layout.fragment_main, container,
             false)
         instantiateElements(rootView)
         root = rootView
