+++ app/src/main/res/layout/fragment_gameplay.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameplayFragment">
+    tools:context=".views.GameplayFragment">
 
 
     <TextView
